main, service: move vesionbundle to pkg/project (#2052)
