+++ pkg/project/version_bundle.go
@@ -1,11 +1,10 @@
-package versionbundle
+package project
 
 import (
-	"github.com/giantswarm/aws-operator/pkg/project"
 	"github.com/giantswarm/versionbundle"
 )
 
-func New() versionbundle.Bundle {
+func NewBundle() versionbundle.Bundle {
 	return versionbundle.Bundle{
 		Changelogs: []versionbundle.Changelog{
 			{
@@ -51,13 +50,7 @@ func New() versionbundle.Bundle {
 				Version: "1.15.5",
 			},
 		},
-		Name:    "aws-operator",
-		Version: project.BundleVersion(),
-	}
-}
-
-func NewSlice() []versionbundle.Bundle {
-	return []versionbundle.Bundle{
-		New(),
+		Name:    Name(),
+		Version: BundleVersion(),
 	}
 }
